+++ fabric/src/main/java/com/github/phylogeny/bettercushionplacement/datagen/DataGeneration.java
@@ -8,5 +8,6 @@ public class DataGeneration implements DataGeneratorEntrypoint {
     public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
         FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
         pack.addProvider(ModLanguageProvider::new);
+        pack.addProvider(ModBlockTagsProvider::new);
     }
 }